BEH/oscina/requirements.txt
numpy==1.18.1
PyYAML==6.0
scikit-image==0.17.2
scipy==1.4.1
statsmodels==0.11.0
